Delay only auctions by default
