+++ src/prebid.js
@@ -50,6 +50,7 @@ import {getHighestCpm} from './utils/reducers.js';
 import {ORTB_VIDEO_PARAMS, fillVideoDefaults, validateOrtbVideoFields} from './video.js';
 import { ORTB_BANNER_PARAMS } from './banner.js';
 import { BANNER, VIDEO } from './mediaTypes.js';
+import {delayIfPrerendering} from './utils/prerendering.js';
 
 const pbjsInstance = getGlobal();
 const { triggerUserSyncs } = userSync;
@@ -569,12 +570,9 @@ pbjsInstance.requestBids = (function() {
     })
   }, 'requestBids');
 
-  return wrapHook(delegate, delayIfPrerendering(function requestBids(req = {}) {
+  return wrapHook(delegate, delayIfPrerendering(() => !config.getConfig('allowPrerendering'), function requestBids(req = {}) {
     // unlike the main body of `delegate`, this runs before any other hook has a chance to;
     // it's also not restricted in its return value in the way `async` hooks are.
-
-    // Note that we delay this when prerendering because in some configurations requestBids may happen outside of
-    // pbjs.que (e.g. NPM consumers).
 
     // if the request does not specify adUnits, clone the global adUnit array;
     // otherwise, if the caller goes on to use addAdUnits/removeAdUnits, any asynchronous logic
@@ -1023,35 +1021,15 @@ function processQueue(queue) {
 }
 
 /**
- * Returns a wrapper around fn that delays execution until the page if activated, if it was prerendered.
- * https://developer.chrome.com/docs/web-platform/prerender-pages
- */
-function delayIfPrerendering(fn) {
-  return function () {
-    if (document.prerendering && !(config.getConfig('allowPrerendering') || getGlobal().allowPrerendering)) {
-      const that = this;
-      const args = Array.from(arguments);
-      return new Promise((resolve) => {
-        document.addEventListener('prerenderingchange', () => {
-          logInfo(`Auctions are suspended while page is prerendering. Set $$PREBID_GLOBAL$$.allowPrerendering = true to allow auctions during prerendering.`)
-          resolve(fn.apply(that, args))
-        }, {once: true})
-      })
-    } else {
-      return Promise.resolve(fn.apply(this, arguments));
-    }
-  }
-}
-/**
  * @alias module:pbjs.processQueue
  */
-pbjsInstance.processQueue = delayIfPrerendering(function () {
+pbjsInstance.processQueue = function () {
   pbjsInstance.que.push = pbjsInstance.cmd.push = quePush;
   insertLocatorFrame();
   hook.ready();
   processQueue(pbjsInstance.que);
   processQueue(pbjsInstance.cmd);
-});
+};
 
 /**
  * @alias module:pbjs.triggerBilling
